apply builder pattern
