+++ formatter/builder.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 	"strings"
 
+	"github.com/fatih/color"
 	"github.com/gnoswap-labs/lint/internal"
 	tt "github.com/gnoswap-labs/lint/internal/types"
 )
@@ -18,128 +19,219 @@ const (
 	SliceBound          = "slice-bounds-check"
 )
 
+const tabWidth = 8
+
+var (
+	errorStyle      = color.New(color.FgRed, color.Bold)
+	warningStyle    = color.New(color.FgHiYellow, color.Bold)
+	ruleStyle       = color.New(color.FgYellow, color.Bold)
+	fileStyle       = color.New(color.FgCyan, color.Bold)
+	lineStyle       = color.New(color.FgBlue, color.Bold)
+	messageStyle    = color.New(color.FgRed, color.Bold)
+	suggestionStyle = color.New(color.FgGreen, color.Bold)
+)
+
+// IssueFormatter is the interface that wraps the Format method.
+// Implementations of this interface are responsible for formatting specific types of lint issues.
+type IssueFormatter interface {
+	Format(issue tt.Issue, snippet *internal.SourceCode) string
+}
+
+// GenetateFormattedIssue formats a slice of issues into a human-readable string.
+// It uses the appropriate formatter for each issue based on its rule.
+func GenetateFormattedIssue(issues []tt.Issue, snippet *internal.SourceCode) string {
+	var builder strings.Builder
+	for _, issue := range issues {
+		builder.WriteString(formatIssueHeader(issue))
+		formatter := getFormatter(issue.Rule)
+		builder.WriteString(formatter.Format(issue, snippet))
+	}
+	return builder.String()
+}
+
+// getFormatter is a factory function that returns the appropriate IssueFormatter
+// based on the given rule.
+// If no specific formatter is found for the given rule, it returns a GeneralIssueFormatter.
+func getFormatter(rule string) IssueFormatter {
+	switch rule {
+	case UnnecessaryElse:
+		return &UnnecessaryElseFormatter{}
+	case SimplifySliceExpr:
+		return &SimplifySliceExpressionFormatter{}
+	case UnnecessaryTypeConv:
+		return &UnnecessaryTypeConversionFormatter{}
+	case CycloComplexity:
+		return &CyclomaticComplexityFormatter{}
+	case EmitFormat:
+		return &EmitFormatFormatter{}
+	case SliceBound:
+		return &SliceBoundsCheckFormatter{}
+	default:
+		return &GeneralIssueFormatter{}
+	}
+}
+
+// formatIssueHeader creates a formatted header string for a given issue.
+// The header includes the rule and the filename. (e.g. "error: unused-variable\n --> test.go")
+func formatIssueHeader(issue tt.Issue) string {
+	return errorStyle.Sprint("error: ") + ruleStyle.Sprint(issue.Rule) + "\n" +
+		lineStyle.Sprint(" --> ") + fileStyle.Sprint(issue.Filename) + "\n"
+}
+
+func buildSuggestion(result *strings.Builder, issue tt.Issue, lineStyle, suggestionStyle *color.Color, startLine int) {
+	maxLineNumWidth := calculateMaxLineNumWidth(issue.End.Line)
+	padding := strings.Repeat(" ", maxLineNumWidth)
+
+	result.WriteString(suggestionStyle.Sprintf("Suggestion:\n"))
+	for i, line := range strings.Split(issue.Suggestion, "\n") {
+		lineNum := fmt.Sprintf("%d", startLine+i)
+
+		if maxLineNumWidth < len(lineNum) {
+			maxLineNumWidth = len(lineNum)
+		}
+
+		result.WriteString(lineStyle.Sprintf("%s%s | ", padding[:maxLineNumWidth-len(lineNum)], lineNum))
+		result.WriteString(fmt.Sprintf("%s\n", line))
+	}
+	result.WriteString("\n")
+}
+
+func buildNote(result *strings.Builder, issue tt.Issue, suggestionStyle *color.Color) {
+	result.WriteString(suggestionStyle.Sprint("Note: "))
+	result.WriteString(fmt.Sprintf("%s\n", issue.Note))
+	result.WriteString("\n")
+}
+
+/***** Issue Formatter Builder *****/
+
 type IssueFormatterBuilder struct {
-	result strings.Builder
-	issue tt.Issue
+	result  strings.Builder
+	issue   tt.Issue
 	snippet *internal.SourceCode
 }
 
 func NewIssueFormatterBuilder(issue tt.Issue, snippet *internal.SourceCode) *IssueFormatterBuilder {
-    return &IssueFormatterBuilder{
-        issue:   issue,
-        snippet: snippet,
-    }
+	return &IssueFormatterBuilder{
+		issue:   issue,
+		snippet: snippet,
+	}
 }
 
 func (b *IssueFormatterBuilder) AddHeader() *IssueFormatterBuilder {
 	// add error type and rule name
-    b.result.WriteString(errorStyle.Sprint("error: "))
-    b.result.WriteString(ruleStyle.Sprintln(b.issue.Rule))
+	b.result.WriteString(errorStyle.Sprint("error: "))
+	b.result.WriteString(ruleStyle.Sprintln(b.issue.Rule))
 
 	// add file name
-    b.result.WriteString(lineStyle.Sprint(" --> "))
-    b.result.WriteString(fileStyle.Sprintln(b.issue.Filename))
+	b.result.WriteString(lineStyle.Sprint(" --> "))
+	b.result.WriteString(fileStyle.Sprintln(b.issue.Filename))
 
 	// add separator
-    maxLineNumWidth := calculateMaxLineNumWidth(b.issue.End.Line)
-    padding := strings.Repeat(" ", maxLineNumWidth+1)
-    b.result.WriteString(lineStyle.Sprintf("%s|\n", padding))
+	maxLineNumWidth := calculateMaxLineNumWidth(b.issue.End.Line)
+	padding := strings.Repeat(" ", maxLineNumWidth+1)
+	b.result.WriteString(lineStyle.Sprintf("%s|\n", padding))
 
-    return b
+	return b
 }
 
 func (b *IssueFormatterBuilder) AddCodeSnippet() *IssueFormatterBuilder {
-    startLine := b.issue.Start.Line
-    endLine := b.issue.End.Line
-    maxLineNumWidth := calculateMaxLineNumWidth(endLine)
+	startLine := b.issue.Start.Line
+	endLine := b.issue.End.Line
+	maxLineNumWidth := calculateMaxLineNumWidth(endLine)
 
-    for i := startLine; i <= endLine; i++ {
+	// add separator
+	padding := strings.Repeat(" ", maxLineNumWidth+1)
+	b.result.WriteString(lineStyle.Sprintf("%s|\n", padding))
+
+	for i := startLine; i <= endLine; i++ {
 		// check that the line number does not go out of range of snippet.Lines
-        if i-1 < 0 || i-1 >= len(b.snippet.Lines) {
-            continue
-        }
+		if i-1 < 0 || i-1 >= len(b.snippet.Lines) {
+			continue
+		}
 
-        line := expandTabs(b.snippet.Lines[i-1])
-        lineNum := fmt.Sprintf("%*d", maxLineNumWidth, i)
-        
-        b.result.WriteString(lineStyle.Sprintf("%s | ", lineNum))
-        b.result.WriteString(line + "\n")
-    }
+		line := expandTabs(b.snippet.Lines[i-1])
+		lineNum := fmt.Sprintf("%*d", maxLineNumWidth, i)
 
-    b.result.WriteString("\n")
+		b.result.WriteString(lineStyle.Sprintf("%s | ", lineNum))
+		b.result.WriteString(line + "\n")
+	}
 
-    return b
+	return b
 }
 
 func (b *IssueFormatterBuilder) AddUnderlineAndMessage() *IssueFormatterBuilder {
-    startLine := b.issue.Start.Line
-    endLine := b.issue.End.Line
-    maxLineNumWidth := calculateMaxLineNumWidth(endLine)
-    padding := strings.Repeat(" ", maxLineNumWidth+1)
+	startLine := b.issue.Start.Line
+	endLine := b.issue.End.Line
+	maxLineNumWidth := calculateMaxLineNumWidth(endLine)
+	padding := strings.Repeat(" ", maxLineNumWidth+1)
 
-    b.result.WriteString(lineStyle.Sprintf("%s| ", padding))
+	b.result.WriteString(lineStyle.Sprintf("%s| ", padding))
+
+	if startLine <= 0 || startLine > len(b.snippet.Lines) || endLine <= 0 || endLine > len(b.snippet.Lines) {
+		b.result.WriteString(messageStyle.Sprintf("Error: Invalid line numbers\n"))
+		return b
+	}
 
 	// draw underline from start column to end column
-    underlineStart := calculateVisualColumn(b.snippet.Lines[startLine-1], b.issue.Start.Column)
-    underlineEnd := calculateVisualColumn(b.snippet.Lines[endLine-1], b.issue.End.Column)
-    underlineLength := underlineEnd - underlineStart + 1
-    
-    b.result.WriteString(strings.Repeat(" ", underlineStart))
-    b.result.WriteString(messageStyle.Sprintf("%s\n", strings.Repeat("~", underlineLength)))
+	underlineStart := calculateVisualColumn(b.snippet.Lines[startLine-1], b.issue.Start.Column)
+	underlineEnd := calculateVisualColumn(b.snippet.Lines[endLine-1], b.issue.End.Column)
+	underlineLength := underlineEnd - underlineStart + 1
 
-    b.result.WriteString(lineStyle.Sprintf("%s| ", padding))
-    b.result.WriteString(messageStyle.Sprintf("%s\n\n", b.issue.Message))
+	b.result.WriteString(strings.Repeat(" ", underlineStart))
+	b.result.WriteString(messageStyle.Sprintf("%s\n", strings.Repeat("~", underlineLength)))
 
-    return b
+	b.result.WriteString(lineStyle.Sprintf("%s| ", padding))
+	b.result.WriteString(messageStyle.Sprintf("%s\n\n", b.issue.Message))
+
+	return b
 }
 
 func (b *IssueFormatterBuilder) AddSuggestion() *IssueFormatterBuilder {
-    if b.issue.Suggestion == "" {
-        return b
-    }
+	if b.issue.Suggestion == "" {
+		return b
+	}
 
-    maxLineNumWidth := calculateMaxLineNumWidth(b.issue.End.Line)
-    padding := strings.Repeat(" ", maxLineNumWidth+1)
+	maxLineNumWidth := calculateMaxLineNumWidth(b.issue.End.Line)
+	padding := strings.Repeat(" ", maxLineNumWidth+1)
 
-    b.result.WriteString(suggestionStyle.Sprint("Suggestion:\n"))
-    b.result.WriteString(lineStyle.Sprintf("%s|\n", padding))
+	b.result.WriteString(suggestionStyle.Sprint("Suggestion:\n"))
+	b.result.WriteString(lineStyle.Sprintf("%s|\n", padding))
 
-    // 제안사항을 여러 줄로 나누어 처리
-    suggestionLines := strings.Split(b.issue.Suggestion, "\n")
-    for i, line := range suggestionLines {
-        lineNum := fmt.Sprintf("%*d", maxLineNumWidth, b.issue.Start.Line+i)
-        b.result.WriteString(lineStyle.Sprintf("%s | %s\n", lineNum, line))
-    }
+	suggestionLines := strings.Split(b.issue.Suggestion, "\n")
+	for i, line := range suggestionLines {
+		lineNum := fmt.Sprintf("%*d", maxLineNumWidth, b.issue.Start.Line+i)
+		b.result.WriteString(lineStyle.Sprintf("%s | %s\n", lineNum, line))
+	}
 
-    b.result.WriteString(lineStyle.Sprintf("%s|\n", padding))
-    b.result.WriteString("\n") // 가독성을 위한 빈 줄 추가
+	b.result.WriteString(lineStyle.Sprintf("%s|\n", padding))
+	b.result.WriteString("\n")
 
-    return b
+	return b
 }
 
 func (b *IssueFormatterBuilder) AddNote() *IssueFormatterBuilder {
-    if b.issue.Note == "" {
-        return b
-    }
+	if b.issue.Note == "" {
+		return b
+	}
 
-    b.result.WriteString(suggestionStyle.Sprint("Note: "))
-    b.result.WriteString(b.issue.Note)
-    b.result.WriteString("\n\n")
+	b.result.WriteString(suggestionStyle.Sprint("Note: "))
+	b.result.WriteString(b.issue.Note)
+	b.result.WriteString("\n\n")
 
-    return b
+	return b
 }
 
 type BaseFormatter struct{}
 
 func (f *BaseFormatter) Format(issue tt.Issue, snippet *internal.SourceCode) string {
-    builder := NewIssueFormatterBuilder(issue, snippet)
-    return builder.
-        AddHeader().
-        AddCodeSnippet().
-        AddUnderlineAndMessage().
-        AddSuggestion().
-        AddNote().
-        Build()
+	builder := NewIssueFormatterBuilder(issue, snippet)
+	return builder.
+		AddHeader().
+		AddCodeSnippet().
+		AddUnderlineAndMessage().
+		AddSuggestion().
+		AddNote().
+		Build()
 }
 
 func (b *IssueFormatterBuilder) Build() string {
@@ -168,6 +260,9 @@ func expandTabs(line string) string {
 // calculateVisualColumn calculates the visual column position
 // in a string. taking into account tab characters.
 func calculateVisualColumn(line string, column int) int {
+	if column < 0 {
+		return 0
+	}
 	visualColumn := 0
 	for i, ch := range line {
 		if i+1 == column {
@@ -180,4 +275,14 @@ func calculateVisualColumn(line string, column int) int {
 		}
 	}
 	return visualColumn
+}
+
+func calculateMaxLineLength(lines []string, start, end int) int {
+	maxLen := 0
+	for i := start - 1; i < end; i++ {
+		if len(lines[i]) > maxLen {
+			maxLen = len(lines[i])
+		}
+	}
+	return maxLen
 }
