+++ formatter/slice_bound.go
@@ -1,9 +1,6 @@
 package formatter
 
 import (
-	"fmt"
-	"strings"
-
 	"github.com/gnoswap-labs/lint/internal"
 	tt "github.com/gnoswap-labs/lint/internal/types"
 )
@@ -14,30 +11,22 @@ func (f *SliceBoundsCheckFormatter) Format(
 	issue tt.Issue,
 	snippet *internal.SourceCode,
 ) string {
-	var result strings.Builder
+	builder := NewIssueFormatterBuilder(issue, snippet)
+	return builder.
+		// AddHeader().
+		AddCodeSnippet().
+		AddUnderlineAndMessage().
+		AddWarning().
+		Build()
+}
 
-	maxLineNumWidth := calculateMaxLineNumWidth(issue.End.Line)
-	padding := strings.Repeat(" ", maxLineNumWidth+1)
-
-	startLine := issue.Start.Line
-	endLine := issue.End.Line
-	for i := startLine; i <= endLine; i++ {
-		line := expandTabs(snippet.Lines[i-1])
-		result.WriteString(lineStyle.Sprintf("%s%d | ", padding[:maxLineNumWidth-len(fmt.Sprintf("%d", i))], i))
-		result.WriteString(line + "\n")
+func (b *IssueFormatterBuilder) AddWarning() *IssueFormatterBuilder {
+	b.result.WriteString(warningStyle.Sprint("warning: "))
+	if b.issue.Category == "index-access" {
+		b.result.WriteString("Index access without bounds checking can lead to runtime panics.\n")
+	} else if b.issue.Category == "slice-expression" {
+		b.result.WriteString("Slice expressions without proper length checks may cause unexpected behavior.\n\n")
 	}
 
-	result.WriteString(lineStyle.Sprintf("%s| ", padding))
-	result.WriteString(messageStyle.Sprintf("%s\n", strings.Repeat("~", calculateMaxLineLength(snippet.Lines, startLine, endLine))))
-	result.WriteString(lineStyle.Sprintf("%s| ", padding))
-	result.WriteString(messageStyle.Sprintf("%s\n\n", issue.Message))
-
-	result.WriteString(warningStyle.Sprint("warning: "))
-	if issue.Category == "index-access" {
-		result.WriteString("Index access without bounds checking can lead to runtime panics.\n")
-	} else if issue.Category == "slice-expression" {
-		result.WriteString("Slice expressions without proper length checks may cause unexpected behavior.\n\n")
-	}
-
-	return result.String()
+	return b
 }
