+++ formatter/unnecessary_type_conv.go
@@ -1,9 +1,6 @@
 package formatter
 
 import (
-	"fmt"
-	"strings"
-
 	"github.com/gnoswap-labs/lint/internal"
 	tt "github.com/gnoswap-labs/lint/internal/types"
 )
@@ -11,37 +8,12 @@ import (
 type UnnecessaryTypeConversionFormatter struct{}
 
 func (f *UnnecessaryTypeConversionFormatter) Format(issue tt.Issue, snippet *internal.SourceCode) string {
-	var result strings.Builder
-
-	// 1. Calculate dimensions
-	startLine := issue.Start.Line
-	endLine := issue.Start.Line // Only one line for type conversion
-	maxLineNumWidth := calculateMaxLineNumWidth(endLine)
-
-	// 2. Write header
-	padding := strings.Repeat(" ", maxLineNumWidth+1)
-	result.WriteString(lineStyle.Sprintf("%s|\n", padding))
-
-	// 3. Write code snippet
-	line := expandTabs(snippet.Lines[startLine-1])
-	lineNum := fmt.Sprintf("%*d", maxLineNumWidth, startLine)
-	result.WriteString(lineStyle.Sprintf("%s | %s\n", lineNum, line))
-
-	// 4. Write underline and message
-	visualColumn := calculateVisualColumn(line, issue.Start.Column)
-	result.WriteString(lineStyle.Sprintf("%s| ", padding))
-	result.WriteString(strings.Repeat(" ", visualColumn))
-	result.WriteString(messageStyle.Sprintf("^ %s\n\n", issue.Message))
-
-	// 5. Write suggestion
-	if issue.Suggestion != "" {
-		buildSuggestion(&result, issue, lineStyle, suggestionStyle, startLine)
-	}
-
-	// 6. Write note
-	if issue.Note != "" {
-		buildNote(&result, issue, suggestionStyle)
-	}
-
-	return result.String()
+	builder := NewIssueFormatterBuilder(issue, snippet)
+	return builder.
+		// AddHeader().
+		AddCodeSnippet().
+		AddUnderlineAndMessage().
+		AddSuggestion().
+		AddNote().
+		Build()
 }
